修复评论编辑预览界面还存在编辑按钮的问题
Fixes #104
