+++ lib/pages/board/widgets/comment_list.dart
@@ -1,14 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_markdown/flutter_markdown.dart';
-import 'package:smart_home/blocs/board/blocs.dart';
-
+import 'package:smart_home/blocs/core/blocs.dart';
 import 'package:smart_home/models/board.dart';
 import 'package:smart_home/models/models.dart';
-import 'package:smart_home/pages/board/comment_edit_page.dart';
-import 'package:smart_home/pages/board/widgets/item_title.dart';
-import 'package:smart_home/repositories/board_repository.dart';
-import 'package:smart_home/utils/show_snack_bar.dart';
+import 'package:smart_home/pages/board/widgets/comment_item.dart';
 
 class SliverCommentList extends StatelessWidget {
   final List<Comment> comments;
@@ -20,95 +15,17 @@ class SliverCommentList extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final loginUser =
+        context.select((AppPreferencesBloc b) => b.state.loginUser);
     return SliverList(
       delegate: SliverChildBuilderDelegate(
         (BuildContext context, int index) {
-          return CommentItem(comment: comments[index]);
+          return CommentItem(
+            comment: comments[index],
+            showMenu: loginUser == comments[index].user,
+          );
         },
         childCount: comments.length,
-      ),
-    );
-  }
-}
-
-class CommentItem extends StatelessWidget {
-  final Comment comment;
-
-  const CommentItem({
-    Key key,
-    @required this.comment,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      key: ValueKey(comment.id),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.start,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          ItemTitle(
-            user: comment.user,
-            editedAt: comment.editedAt,
-            onSelected: (Menu menu) async {
-              switch (menu) {
-                case Menu.delete:
-                  showDialog(
-                    context: context,
-                    builder: (_) => AlertDialog(
-                      title: Text('删除评论'),
-                      content: Text('你确认要删除该评论？'),
-                      actions: <Widget>[
-                        FlatButton(
-                          child: Text('否'),
-                          onPressed: () {
-                            Navigator.of(context).pop();
-                          },
-                        ),
-                        FlatButton(
-                          child: Text('是'),
-                          onPressed: () {
-                            showInfoSnackBar('正在删除...', duration: 1);
-                            BlocProvider.of<CommentEditBloc>(context)
-                                .add(CommentDeleted(comment: comment));
-                            Navigator.pop(context);
-                          },
-                        ),
-                      ],
-                    ),
-                  );
-                  break;
-                case Menu.edit:
-                  await Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (_) => BlocProvider(
-                        create: (context) => CommentEditBloc(
-                            boardRepository:
-                                RepositoryProvider.of<BoardRepository>(
-                                    context)),
-                        child: CommentEditPage(
-                          isEditing: true,
-                          comment: comment,
-                        ),
-                      ),
-                    ),
-                  );
-                  BlocProvider.of<TopicDetailBloc>(context)
-                      .add(TopicDetailRefreshed());
-                  break;
-              }
-            },
-          ),
-          Padding(
-            padding: const EdgeInsets.fromLTRB(16, 0, 16, 0),
-            child: MarkdownBody(
-              data: comment.body,
-              selectable: true,
-            ),
-          ),
-          Divider(),
-        ],
       ),
     );
   }
